Refactor utils file structure
