--- app/utils/config.toml
@@ -1,110 +1,0 @@
-# Poricom Default Configuration File
-# 
-# Copyright (C) `2021-2022` `<Alarcon Ace Belen>`
-# 
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-# 
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-# 
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-# Lists
-IMAGE_EXTENSIONS = [ "*.bmp", "*.gif", "*.jpg", "*.jpeg", "*.png", "*.pbm", "*.pgm", "*.ppm", "*.webp", "*.xbm", "*.xpm",]
-LANGUAGE = [ " Japanese", " Korean", " Chinese SIM", " Chinese TRA  ", " English",]
-ORIENTATION = [ " Vertical", " Horizontal",]
-FONT_STYLE = [ " Helvetica", " Poppins", " Arial", " Verdana", " Times New Roman",]
-FONT_SIZE = [ " 12", " 14", " 16", " 20", " 24", " 32", " 40", " 56", " 72",]
-IMAGE_SCALING= [ " Fit to Width", " Fit to Height", " Fit to Screen",]
-MODIFIER = [ " Ctrl", " Shift", " Alt", " Ctrl+Alt", " Shift+Alt", " Shift+Ctrl", " Shift+Alt+Ctrl", " No Modifier"]
-
-# Filepath
-STYLES_PATH = "./assets/"
-NAV_ROOT = "./assets/images/"
-TBAR_ICONS = "./assets/images/icons/"
-TBAR_ICONS_LIGHT = "./assets/images/icons/"
-LANG_PATH = "./assets/languages/"
-
-STYLES_DEFAULT = "./assets/styles.qss"
-LOGO = "./assets/images/icons/logo.ico"
-HOME_IMAGE = "./assets/images/home.png"
-ABOUT_IMAGE = "./assets/images/about.png"
-TBAR_ICON_DEFAULT = "./assets/images/icons/default_icon.png"
-
-# Sizes
-RBN_HEIGHT = 2.4
-TBAR_ISIZE_REL = 0.1
-TBAR_ISIZE_MARGIN = 1.3
-NAV_VIEW_RATIO = [ 3, 11,]
-
-# Mode
-VIEW_IMAGE_MODE = 0
-SPLIT_VIEW_MODE = false
-
-[SELECTED_INDEX]
-language = 0
-orientation = 0
-fontStyle = 0
-fontSize = 2
-imageScaling = 0
-modifier = 2
-
-[PICKER_INDEX]
-language = 49
-orientation = 50
-fontStyle = 51
-fontSize = 52
-imageScaling = 53
-modifier = 54
-
-[SHORTCUT]
-captureExternal = "Alt+Q"
-captureExternalKey = "Q"
-captureExternalTip = "This will minimize the app and perform OCR on the current screen. Alternatively, you may use the shortcut: "
-
-# Ribbon buttons
-[TBAR_FUNCS]
-
-  [TBAR_FUNCS.SETTINGS]
-
-    [TBAR_FUNCS.SETTINGS.captureExternalHelper]
-    helpTitle = "External capture"
-    helpMsg = "This will minimize the app and perform OCR on the current screen. Alternatively, you may use the shortcut Alt+Q."
-    path = "captureExternalHelper.png"
-    toggle = false
-    align = "AlignLeft"
-    iconH = 1.0
-    iconW = 1.0
-
-    [TBAR_FUNCS.SETTINGS.modifyFontSettings]
-    helpTitle = "Modify preview text"
-    helpMsg = "Change font style and font size of preview text."
-    path = "modifyFontSettings.png"
-    toggle = false
-    align = "AlignLeft"
-    iconH = 1.0
-    iconW = 1.0
-
-    [TBAR_FUNCS.SETTINGS.modifyHotkeys]
-    helpTitle = "Remap hotkeys"
-    helpMsg = "Change shortcut for external captures."
-    path = "modifyHotkeys.png"
-    toggle = false
-    align = "AlignLeft"
-    iconH = 1.0
-    iconW = 1.0
-
-    [TBAR_FUNCS.SETTINGS.toggleLogging]
-    helpTitle = "Enable text logging"
-    helpMsg = "Save detected text to a text file located in the current project directory."
-    path = "toggleLogging.png"
-    toggle = true
-    align = "AlignLeft"
-    iconH = 1.0
-    iconW = 1.0
